+++ src/commands/__init__.py
@@ -2,7 +2,8 @@ from ...bootstrap import reload
 
 reload("src.commands", ["package_report", "override_report", "diff_report",
        "diff_single", "toggle_override", "diff_override", "edit_override",
-       "delete_override", "freshen_override", "modify_mark"])
+       "delete_override", "freshen_override", "diff_package",
+       "freshen_package", "modify_mark"])
 
 from .package_report import OverrideAuditPackageReportCommand
 from .override_report import OverrideAuditOverrideReportCommand
@@ -13,14 +14,15 @@ from .diff_override import OverrideAuditDiffOverrideCommand
 from .edit_override import OverrideAuditEditOverrideCommand
 from .delete_override import OverrideAuditDeleteOverrideCommand
 from .freshen_override import OverrideAuditFreshenOverrideCommand
+from .diff_package import OverrideAuditDiffPackageCommand
+from .freshen_package import OverrideAuditFreshenPackageCommand
 from .modify_mark import OverrideAuditModifyMarkCommand
 
 __all__ = [
-    # Reports
+    # Report generation commands
     "OverrideAuditPackageReportCommand",
     "OverrideAuditOverrideReportCommand",
     "OverrideAuditDiffReportCommand",
-    "OverrideAuditDiffSingleCommand",
 
     # Override commands
     "OverrideAuditToggleOverrideCommand",
@@ -29,6 +31,11 @@ __all__ = [
     "OverrideAuditDeleteOverrideCommand",
     "OverrideAuditFreshenOverrideCommand",
 
+    # Package commands
+    "OverrideAuditDiffPackageCommand",
+    "OverrideAuditFreshenPackageCommand",
+
     # General
+    "OverrideAuditDiffSingleCommand",
     "OverrideAuditModifyMarkCommand"
 ]